Struct magick_rust::MagickWand

pub struct MagickWand {
    pub wand: *mut MagickWand,
}

Fields
wand: *mut MagickWand
Methods

impl MagickWand

pub fn new() -> Self

pub fn clear_exception(&mut self) -> Result<(), &'static str>

pub fn get_exception_type(&self) -> ExceptionType

pub fn get_exception(&self) -> Result<(String, ExceptionType), &'static str>

pub fn is_wand(&self) -> Result<(), &'static str>

impl MagickWand

MagickWand is a Rustic wrapper to the Rust bindings to ImageMagick.

Instantiating a MagickWand will construct an ImageMagick "wand" on which operations can be performed via the MagickWand functions. When the MagickWand is dropped, the ImageMagick wand will be destroyed as well.

pub fn new_image(
&self,
columns: size_t,
rows: size_t,
pixel_wand: &PixelWand
) -> Result<(), &'static str>

pub fn set_option(&mut self, key: &str, value: &str) -> Result<(), &'static str>

pub fn annotate_image(
&mut self,
drawing_wand: &DrawingWand,
x: f64,
y: f64,
angle: f64,
text: &str
) -> Result<(), &'static str>

pub fn add_image(&mut self, other_wand: &MagickWand) -> Result<(), &'static str>

Add all images from another wand to this wand at the current index.

pub fn append_all(&mut self, stack: bool) -> MagickWand

pub fn label_image(&self, label: &str) -> Result<(), &'static str>

pub fn write_images(&self, path: &str, adjoin: bool) -> Result<(), &'static str>

pub fn read_image(&self, path: &str) -> Result<(), &'static str>

Read the image data from the named file.

pub fn read_image_blob<T: AsRef<[u8]>>(
&self,
data: T
) -> Result<(), &'static str>

Read the image data from the vector of bytes.

pub fn ping_image(&self, path: &str) -> Result<(), &'static str>

Same as read_image, but reads only the width, height, size and format of an image, without reading data.

pub fn ping_image_blob<T: AsRef<[u8]>>(
&self,
data: T
) -> Result<(), &'static str>

Same as read_image, but reads only the width, height, size and format of an image, without reading data.

pub fn compare_images(
&self,
reference: &MagickWand,
metric: MetricType
) -> (f64, Option<MagickWand>)

Compare two images and return tuple (distortion, diffImage) diffImage is None if distortion == 0

pub fn compose_images(
&self,
reference: &MagickWand,
composition_operator: CompositeOperator,
clip_to_self: bool,
x: isize,
y: isize
) -> Result<(), &'static str>

Compose another image onto self at (x, y) using composition_operator

pub fn clut_image(
&self,
clut_wand: &MagickWand,
method: PixelInterpolateMethod
) -> Result<(), &'static str>

pub fn set_size(
&self,
columns: size_t,
rows: size_t
) -> Result<(), &'static str>

pub fn extend_image(
&self,
width: usize,
height: usize,
x: isize,
y: isize
) -> Result<(), &'static str>

Extend the image as defined by the geometry, gravity, and wand background color. Set the (x,y) offset of the geometry to move the original wand relative to the extended wand.

pub fn profile_image<'a, T: Into<Option<&'a [u8]>>>(
&self,
name: &str,
profile: T
) -> Result<(), &'static str>

pub fn flip_image(&self) -> Result<(), &'static str>

pub fn flop_image(&self) -> Result<(), &'static str>

pub fn gaussian_blur_image(
&self,
radius: f64,
sigma: f64
) -> Result<(), &'static str>

pub fn adaptive_resize_image(
&self,
width: usize,
height: usize
) -> Result<(), &'static str>

Adaptively resize the currently selected image.

pub fn rotate_image(
&self,
background: &PixelWand,
degrees: f64
) -> Result<(), &'static str>

Rotate the currently selected image by the given number of degrees, filling any empty space with the background color of a given PixelWand

pub fn trim_image(&self, fuzz: f64) -> Result<(), &'static str>

Trim the image removing the backround color from the edges.

pub fn get_image_width(&self) -> usize

Retrieve the width of the image.

pub fn get_image_height(&self) -> usize

Retrieve the height of the image.

pub fn get_image_page(&self) -> (usize, usize, isize, isize)

Retrieve the page geometry (width, height, x offset, y offset) of the image.

pub fn reset_image_page(&self, page_geometry: &str) -> Result<(), &'static str>

Reset the Wand page canvas and position.

pub fn get_image_property(&self, name: &str) -> Result<String, &'static str>

Retrieve the named image property value.

pub fn set_image_property(
&self,
name: &str,
value: &str
) -> Result<(), &'static str>

Set the named image property.

pub fn get_image_pixel_color(&self, x: isize, y: isize) -> Option<PixelWand>

Returns a PixelWand instance for the pixel specified by x and y offests.

pub fn set_sampling_factors(
&self,
samplingFactors: &[f64]
) -> Result<(), &'static str>

Sets the image sampling factors.

samplingFactors: An array of floats representing the sampling factor for each color component (in RGB order).

pub fn get_image_histogram(&self) -> Option<Vec<PixelWand>>

Returns the image histogram as a vector of PixelWand instances for every unique color.

pub fn sharpen_image(&self, radius: f64, sigma: f64) -> Result<(), &'static str>

Sharpens an image. We convolve the image with a Gaussian operator of the given radius and standard deviation (sigma). For reasonable results, the radius should be larger than sigma. Use a radius of 0 and SharpenImage() selects a suitable radius for you.

radius: the radius of the Gaussian, in pixels, not counting the center pixel.

sigma: the standard deviation of the Gaussian, in pixels.

pub fn set_background_color(
&self,
pixel_wand: &PixelWand
) -> Result<(), &'static str>

Set the background color.

pub fn set_image_background_color(
&self,
pixel_wand: &PixelWand
) -> Result<(), &'static str>

Set the image background color.

pub fn get_image_resolution(&self) -> Result<(f64, f64), &'static str>

Returns the image resolution as a pair (horizontal resolution, vertical resolution)

pub fn set_image_resolution(
&self,
x_resolution: f64,
y_resolution: f64
) -> Result<(), &'static str>

Sets the image resolution

pub fn set_resolution(
&self,
x_resolution: f64,
y_resolution: f64
) -> Result<(), &'static str>

Sets the wand resolution

pub fn sepia_tone_image(&self, threshold: f64) -> Result<(), &'static str>

Returns the image resolution as a pair (horizontal resolution, vertical resolution)

pub fn export_image_pixels(
&self,
x: isize,
y: isize,
width: usize,
height: usize,
map: &str
) -> Option<Vec<u8>>

Extracts pixel data from the image as a vector of 0..255 values defined by map. See https://www.imagemagick.org/api/magick-image.php#MagickExportImagePixels for more information.

pub fn resize_image(&self, width: usize, height: usize, filter: FilterType)

Resize the image to the specified width and height, using the specified filter type.

pub fn crop_image(
&self,
width: usize,
height: usize,
x: isize,
y: isize
) -> Result<(), &'static str>

Extract a region of the image. The width and height is used as the size of the region. X and Y is the offset.

pub fn resample_image(
&self,
x_resolution: f64,
y_resolution: f64,
filter: FilterType
)

Resample the image to the specified horizontal and vertical resolution, using the specified filter type.

pub fn fit(&self, width: size_t, height: size_t)

Resize the image to fit within the given dimensions, maintaining the current aspect ratio.

pub fn requires_orientation(&self) -> bool

Detect if the loaded image is not in top-left orientation, and hence should be "auto" oriented so it is suitable for viewing.

pub fn auto_orient(&self) -> bool

Automatically adjusts the loaded image so that its orientation is suitable for viewing (i.e. top-left orientation).

Returns true if successful or false if an error occurred.

pub fn write_image(&self, path: &str) -> Result<(), &'static str>

Write the current image to the provided path.

pub fn write_image_blob(&self, format: &str) -> Result<Vec<u8>, &'static str>

Write the image in the desired format to a new blob.

The format argument may be any ImageMagick supported image format (e.g. GIF, JPEG, PNG, etc).

pub fn write_images_blob(&self, format: &str) -> Result<Vec<u8>, &'static str>

Write the images in the desired format to a new blob.

The format argument may be any ImageMagick supported image format (e.g. GIF, JPEG, PNG, etc).

pub fn transform_image_colorspace(
&self,
colorspace: ColorspaceType
) -> Result<(), &'static str>

Set the image colorspace, transforming (unlike set_image_colorspace) image data in the process.

pub fn set_image_alpha_channel(
&self,
alpha_channel: AlphaChannelOption
) -> Result<(), &'static str>

Set the image alpha channel mode.

pub fn quantize_image(
&self,
number_of_colors: size_t,
colorspace: ColorspaceType,
tree_depth: size_t,
dither_method: DitherMethod,
measure_error: MagickBooleanType
) -> Result<(), &'static str>

Reduce the number of colors in the image.

pub fn quantize_images(
&self,
number_of_colors: size_t,
colorspace: ColorspaceType,
tree_depth: size_t,
dither_method: DitherMethod,
measure_error: MagickBooleanType
) -> Result<(), &'static str>

Reduce the number of colors in the image.

pub fn unique_image_colors(&self) -> Result<(), &'static str>

Discard all but one of any pixel color.

pub fn get_image_colors(&self) -> size_t

pub fn get_filename(&self) -> Result<String, &'static str>

pub fn set_filename(&mut self, s: &str) -> Result<(), &'static str>

pub fn get_font(&self) -> Result<String, &'static str>

pub fn set_font(&mut self, s: &str) -> Result<(), &'static str>

pub fn get_format(&self) -> Result<String, &'static str>

pub fn set_format(&mut self, s: &str) -> Result<(), &'static str>

pub fn get_image_filename(&self) -> Result<String, &'static str>

pub fn set_image_filename(&mut self, s: &str) -> Result<(), &'static str>

pub fn get_image_format(&self) -> Result<String, &'static str>

pub fn set_image_format(&mut self, s: &str) -> Result<(), &'static str>

pub fn fmt_string_settings(&self, f: &mut Formatter, prefix: &str) -> Result

pub fn get_colorspace(&self) -> ColorspaceType

pub fn set_colorspace(&mut self, v: ColorspaceType) -> Result<(), &'static str>

pub fn get_compression(&self) -> CompressionType

pub fn set_compression(
&mut self,
v: CompressionType
) -> Result<(), &'static str>

pub fn get_compression_quality(&self) -> size_t

pub fn set_compression_quality(&mut self, v: size_t) -> Result<(), &'static str>

pub fn get_gravity(&self) -> GravityType

pub fn set_gravity(&mut self, v: GravityType) -> Result<(), &'static str>

pub fn get_image_colorspace(&self) -> ColorspaceType

pub fn set_image_colorspace(
&mut self,
v: ColorspaceType
) -> Result<(), &'static str>

pub fn get_image_compose(&self) -> CompositeOperator

pub fn set_image_compose(
&mut self,
v: CompositeOperator
) -> Result<(), &'static str>

pub fn get_image_compression(&self) -> CompressionType

pub fn set_image_compression(
&mut self,
v: CompressionType
) -> Result<(), &'static str>

pub fn get_image_compression_quality(&self) -> size_t

pub fn set_image_compression_quality(
&mut self,
v: size_t
) -> Result<(), &'static str>

pub fn get_image_delay(&self) -> size_t

pub fn set_image_delay(&mut self, v: size_t) -> Result<(), &'static str>

pub fn get_image_depth(&self) -> size_t

pub fn set_image_depth(&mut self, v: size_t) -> Result<(), &'static str>

pub fn get_image_dispose(&self) -> DisposeType

pub fn set_image_dispose(&mut self, v: DisposeType) -> Result<(), &'static str>

pub fn get_image_endian(&self) -> EndianType

pub fn set_image_endian(&mut self, v: EndianType) -> Result<(), &'static str>

pub fn get_image_fuzz(&self) -> f64

pub fn set_image_fuzz(&mut self, v: f64) -> Result<(), &'static str>

pub fn get_image_gamma(&self) -> f64

pub fn set_image_gamma(&mut self, v: f64) -> Result<(), &'static str>

pub fn get_image_gravity(&self) -> GravityType

pub fn set_image_gravity(&mut self, v: GravityType) -> Result<(), &'static str>

pub fn get_image_interlace_scheme(&self) -> InterlaceType

pub fn set_image_interlace_scheme(
&mut self,
v: InterlaceType
) -> Result<(), &'static str>

pub fn get_image_interpolate_method(&self) -> PixelInterpolateMethod

pub fn set_image_interpolate_method(
&mut self,
v: PixelInterpolateMethod
) -> Result<(), &'static str>

pub fn get_image_iterations(&self) -> size_t

pub fn set_image_iterations(&mut self, v: size_t) -> Result<(), &'static str>

pub fn get_image_orientation(&self) -> OrientationType

pub fn set_image_orientation(
&mut self,
v: OrientationType
) -> Result<(), &'static str>

pub fn get_image_rendering_intent(&self) -> RenderingIntent

pub fn set_image_rendering_intent(
&mut self,
v: RenderingIntent
) -> Result<(), &'static str>

pub fn get_image_scene(&self) -> size_t

pub fn set_image_scene(&mut self, v: size_t) -> Result<(), &'static str>

pub fn get_image_type(&self) -> ImageType

pub fn set_image_type(&mut self, v: ImageType) -> Result<(), &'static str>

pub fn get_image_units(&self) -> ResolutionType

pub fn set_image_units(&mut self, v: ResolutionType) -> Result<(), &'static str>

pub fn get_interlace_scheme(&self) -> InterlaceType

pub fn set_interlace_scheme(
&mut self,
v: InterlaceType
) -> Result<(), &'static str>

pub fn get_interpolate_method(&self) -> PixelInterpolateMethod

pub fn set_interpolate_method(
&mut self,
v: PixelInterpolateMethod
) -> Result<(), &'static str>

pub fn get_iterator_index(&self) -> ssize_t

pub fn set_iterator_index(&mut self, v: ssize_t) -> Result<(), &'static str>

pub fn get_orientation(&self) -> OrientationType

pub fn set_orientation(
&mut self,
v: OrientationType
) -> Result<(), &'static str>

pub fn get_pointsize(&self) -> f64

pub fn set_pointsize(&mut self, v: f64) -> Result<(), &'static str>

pub fn get_type(&self) -> ImageType

pub fn set_type(&mut self, v: ImageType) -> Result<(), &'static str>

pub fn fmt_checked_settings(&self, f: &mut Formatter, prefix: &str) -> Result